Enhances email modal and issue details
Improves the email modal component by refactoring the layout for better responsiveness and user experience. It introduces a new SCSS file for styling, replaces the old one, and adjusts the form fields for improved readability.

Enhances the issue detail view by integrating GLightbox for a better photo gallery experience, allowing users to zoom and navigate through images. Adds an "Urgent" chip to issues marked as such, improving visual communication of issue priority.

The location selection component is also updated to focus on Sector 5 as the only selectable option for the MVP version.